Move hydration api utility functions to api directory
